Deleted static web view
