+++ Wikivoyage/LocationWebViewController+Helpers.swift
@@ -10,9 +10,37 @@ import Alamofire
 import SwiftyJSON
 import MagicalRecord
 import MBProgressHUD
+import WebKit
 import MapKit
 
 extension LocationWebViewController {
+    
+    // MARK: - Web
+    
+    func isInternalLink(webView: WKWebView, navigationAction: WKNavigationAction) -> Bool {
+        if let currentURL = webView.URL, requestURL = navigationAction.request.URL {
+            if let currentScheme = currentURL.scheme, currentHost = currentURL.host, currentPathComponents = currentURL.pathComponents as? [String], requestScheme = requestURL.scheme, requestHost = requestURL.host, requestPathComponents = requestURL.pathComponents as? [String] {
+                // An internal link has the same scheme, host and path components
+                if currentScheme == requestScheme && currentHost == requestHost && areArraysEqual(firstArray: currentPathComponents, secondArray: requestPathComponents) {
+                    return true
+                }
+            }
+        }
+        
+        return false
+    }
+    
+    func areArraysEqual(firstArray a: [String], secondArray b: [String]) -> Bool {
+        if a.count != b.count {
+            return false
+        }
+        for i in 0..<a.count {
+            if a[i] != b[i] {
+                return false
+            }
+        }
+        return true
+    }
     
     // MARK: - Favorite
